@@ -1,22 +1,28 @@
-# DRL models from ElegantRL: https://github.com/AI4Finance-Foundation/ElegantRL
+"""
+DRL models from ElegantRL: https://github.com/AI4Finance-Foundation/ElegantRL
+"""
+from __future__ import annotations
+
 import torch
 from elegantrl.agents import AgentDDPG
 from elegantrl.agents import AgentPPO
 from elegantrl.agents import AgentSAC
 from elegantrl.agents import AgentTD3
 from elegantrl.train.config import Arguments
+from elegantrl.train.run import init_agent
+from elegantrl.train.run import train_and_evaluate
+
 # from elegantrl.agents import AgentA2C
-from elegantrl.train.run import train_and_evaluate, init_agent
 
 MODELS = {"ddpg": AgentDDPG, "td3": AgentTD3, "sac": AgentSAC, "ppo": AgentPPO}
 OFF_POLICY_MODELS = ["ddpg", "td3", "sac"]
 ON_POLICY_MODELS = ["ppo"]
-"""MODEL_KWARGS = {x: config.__dict__[f"{x.upper()}_PARAMS"] for x in MODELS.keys()}
-
-NOISE = {
-    "normal": NormalActionNoise,
-    "ornstein_uhlenbeck": OrnsteinUhlenbeckActionNoise,
-}"""
+# MODEL_KWARGS = {x: config.__dict__[f"{x.upper()}_PARAMS"] for x in MODELS.keys()}
+#
+# NOISE = {
+#     "normal": NormalActionNoise,
+#     "ornstein_uhlenbeck": OrnsteinUhlenbeckActionNoise,
+# }
 
 
 class DRLAgent:
@@ -54,7 +60,7 @@ class DRLAgent:
         agent = MODELS[model_name]
         if model_name not in MODELS:
             raise NotImplementedError("NotImplementedError")
-        model = Arguments(agent=agent, env=env)
+        model = Arguments(agent_class=agent, env=env)
         model.if_off_policy = model_name in OFF_POLICY_MODELS
         if model_kwargs is not None:
             try:
@@ -83,7 +89,7 @@ class DRLAgent:
             raise NotImplementedError("NotImplementedError")
         agent = MODELS[model_name]
         environment.env_num = 1
-        args = Arguments(agent=agent, env=environment)
+        args = Arguments(agent_class=agent, env=environment)
         args.cwd = cwd
         args.net_dim = net_dimension
         # load agent
@@ -109,10 +115,10 @@ class DRLAgent:
                 state, reward, done, _ = environment.step(action)
 
                 total_asset = (
-                        environment.amount
-                        + (
-                                environment.price_ary[environment.day] * environment.stocks
-                        ).sum()
+                    environment.amount
+                    + (
+                        environment.price_ary[environment.day] * environment.stocks
+                    ).sum()
                 )
                 episode_total_assets.append(total_asset)
                 episode_return = total_asset / environment.initial_total_asset
