@@ -1,13 +1,13 @@
-
-
 # directory
+from __future__ import annotations
+
 DATA_SAVE_DIR = "datasets"
 TRAINED_MODEL_DIR = "trained_models"
 TENSORBOARD_LOG_DIR = "tensorboard_log"
 RESULTS_DIR = "results"
 
 # date format: '%Y-%m-%d'
-TRAIN_START_DATE = "2014-01-01"
+TRAIN_START_DATE = "2014-01-06"  # bug fix: set Monday right, start date set 2014-01-01 ValueError: all the input array dimensions for the concatenation axis must match exactly, but along dimension 0, the array at index 0 has size 1658 and the array at index 1 has size 1657
 TRAIN_END_DATE = "2020-07-31"
 
 TEST_START_DATE = "2020-08-01"
@@ -54,26 +54,23 @@ ERL_PARAMS = {
     "seed": 312,
     "net_dimension": 512,
     "target_step": 5000,
-    "eval_gap": 30
+    "eval_gap": 30,
+    "eval_times": 64,  # bug fix:KeyError: 'eval_times' line 68, in get_model model.eval_times = model_kwargs["eval_times"]
 }
 RLlib_PARAMS = {"lr": 5e-5, "train_batch_size": 500, "gamma": 0.99}
 
 
 # Possible time zones
-TIME_ZONE_SHANGHAI = 'Asia/Shanghai'  # Hang Seng HSI, SSE, CSI
-TIME_ZONE_USEASTERN = 'US/Eastern'  # Dow, Nasdaq, SP
-TIME_ZONE_PARIS = 'Europe/Paris'  # CAC,
-TIME_ZONE_BERLIN = 'Europe/Berlin'  # DAX, TECDAX, MDAX, SDAX
-TIME_ZONE_JAKARTA = 'Asia/Jakarta'  # LQ45
-TIME_ZONE_SELFDEFINED = 'xxx'  # If neither of the above is your time zone, you should define it, and set USE_TIME_ZONE_SELFDEFINED 1.
+TIME_ZONE_SHANGHAI = "Asia/Shanghai"  # Hang Seng HSI, SSE, CSI
+TIME_ZONE_USEASTERN = "US/Eastern"  # Dow, Nasdaq, SP
+TIME_ZONE_PARIS = "Europe/Paris"  # CAC,
+TIME_ZONE_BERLIN = "Europe/Berlin"  # DAX, TECDAX, MDAX, SDAX
+TIME_ZONE_JAKARTA = "Asia/Jakarta"  # LQ45
+TIME_ZONE_SELFDEFINED = "xxx"  # If neither of the above is your time zone, you should define it, and set USE_TIME_ZONE_SELFDEFINED 1.
 USE_TIME_ZONE_SELFDEFINED = 0  # 0 (default) or 1 (use the self defined)
 
 # parameters for data sources
 ALPACA_API_KEY = "xxx"  # your ALPACA_API_KEY
 ALPACA_API_SECRET = "xxx"  # your ALPACA_API_SECRET
-ALPACA_API_BASE_URL = 'https://paper-api.alpaca.markets'  # alpaca url
-BINANCE_BASE_URL = 'https://data.binance.vision/'  # binance url
-
-
-
-
+ALPACA_API_BASE_URL = "https://paper-api.alpaca.markets"  # alpaca url
+BINANCE_BASE_URL = "https://data.binance.vision/"  # binance url
